Moved cmd/tether/msgs to lib/tether/msgs
