+++ lib/portlayer/attach/communication/interactor.go
@@ -24,8 +24,8 @@ import (
 
 	"golang.org/x/crypto/ssh"
 
-	"github.com/vmware/vic/cmd/tether/msgs"
 	"github.com/vmware/vic/lib/migration/feature"
+	"github.com/vmware/vic/lib/tether/msgs"
 	"github.com/vmware/vic/pkg/trace"
 )
